FavoritesView: code cleanup.
